+++ src/types/member.rs
@@ -2,9 +2,9 @@ use crate::types::{Guild, User};
 
 #[derive(Serialize, Deserialize)]
 pub struct Member {
-    pub user_id: Option<i64>,
+    pub user_id: Option<u128>,
     pub user: Option<User>,
 
-    pub guild_id: Option<i64>,
+    pub guild_id: Option<u128>,
     pub guild: Option<Guild>,
 }
